docs: move Arduino sense guide to use Argon
This moves all samples that were made for the Arduino Sense to work on
the Particle Argon, as students will not be using the Sense board.

Sample source code has been adjusted accordingly and has been tested on
the Particle Boron (which performs identically for the features
implemented). Same SoC

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>